документы плюс , минус сделал
