+++ public/locales/KG/translation.json
@@ -186,7 +186,10 @@
     "informationSecurity": "«Маалыматтык коопсуздук» боюнча билим берүү программасы:",
     "miningMechanic": "«Тоокен жабдууларын жана атайын техниканы тейлөө жана оңдоо боюнча слесарь» боюнча билим берүү программасы:",
     "carRepairMechanic": "«Унаа каражаттарын оңдоо боюнча слесарь» боюнча билим берүү программасы:",
-    "regulationsPlit99": "Нормативдик-укуктук акт: ПЛИТ №99 боюнча жобо"
+    "regulationsPlit99": "Нормативдик-укуктук акт: ПЛИТ №99 боюнча жобо",
+    "selfassessmentreports": "Өзүн-өзү баалоо боюнча отчеттор",
+    "Profession": "«Кесиптик окутуу» сабагы боюнча жумушчу программа",
+    "normativeActsKR": "Кыргыз Республикасынын ченемдик укуктук актылары"
   },
   "docsdescriptions": {
     "webDeveloper": "Веб-иштеп чыгуучу жана мультимедиа интерфейстерди түзүү:",
@@ -202,7 +205,10 @@
     "informationSecurity": "Маалыматтык коопсуздук – маалыматты жана санариптик системаларды коргоо:",
     "miningMechanic": "Тоокен жабдууларын жана атайын техниканы тейлөө жана оңдоо – жер алдында жана үстүндө ишенимдүү техника:",
     "carRepairMechanic": "Унаа каражаттарын оңдоо боюнча слесарь – техниканын ишенимдүү абалда болушу үчүн:",
-    "regulationsPlit99": "ПЛИТ №99 боюнча толук жобо — лицейдеги окуу процессин уюштуруу тартиби."
+    "regulationsPlit99": "ПЛИТ №99 боюнча толук жобо — лицейдеги окуу процессин уюштуруу тартиби.",
+    "initvetprogramselfassessment": "Баштапкы кесиптик билим берүү программалары боюнча өзүн-өзү баалоо",
+    "Profession": "Кесиптик окутуу - ийгиликтүү карьера үчүн көндүмдөр:",
+    "ActsPLIT99": "КР Президентине караштуу МТКЛ №99 ченемдик укуктук актылар"
   },
   "links": {
     "vocationalEducation": "Баштапкы кесиптик билим берүү программасы",
@@ -302,9 +308,40 @@
     "human_society_rp_kg": "РП Адам жана коом кырг.",
     "kyrgyz_language_rp_kg": "Кыргыз тили",
     "kyrgyz_language_umk_kg": "Кыргыз тили УМК",
-     "POPD": "«Өндүрүштүк окутуу» сабагы боюнча жумушчу программа",
+    "itVETprogramSelfAssessment": "IT багытындагы баштапкы кесиптик билим берүү программалары боюнча өзүн-өзү баалоо",
+    "geneduselfassessment": "Орто жалпы билим берүү программасы боюнча өзүн-өзү баалоо",
+    "mechvetselfassessment": "Автотранспорт каражаттарын оңдоо слесары кесиби боюнча баштапкы кесиптик билим берүү программаларынын өзүн-өзү баалоосу",
+    "institutionalselfassessment": "Институционалдык деңгээлдеги өзүн-өзү баалоо",
+    "POPD": "«Өндүрүштүк окутуу» сабагы боюнча жумушчу программа",
     "Security": "«Маалыматтык коопсуздуктун негиздери» сабагы боюнча жумушчу программа",
-    "Profession": "«Кесиптик окутуу» сабагы боюнча жумушчу программа"
+    "Profession": "«Кесиптик окутуу» сабагы боюнча жумушчу программа",
+    "ActsPLIT": "КРнин ченемдик укуктук актысы №99 (ПЛИТ)",
+    "ActsExam": "Экзамен-25",
+    "PresidentialDecree": "КР Президентинин жарлыгы",
+    "GovernmentDecree": "КР Өкмөтүнүн токтому",
+    "CabinetOfMinisters": "КР Министрлер Кабинетинин токтому",
+    "IECRegulations": "Ички эксперттик комиссиянын жобосу",
+    "KyrgyzStateEdDocs2025": "Кыргыз тилиндеги мамлекеттик билим берүү документтери жөнүндө жобо (2025)",
+    "StateStandardTemplate": "Мамлекеттик стандарттын макети",
+    "IGA-25": "ЖМА-25 (Жыйынтыктоочу мамлекеттик аттестация)",
+    "LawsOfTheKR": "КР Мыйзамы «Билим берүү жөнүндө»"
+  },
+  "npa": {
+    "title1": "Кыргыз Республикасынын ченемдик укуктук актылар",
+    "title2": "КР Президент алдындагы МТКЛ №99 ченемдик укуктук актылар"
+  },
+  "npa99": {
+    "title1": "ПЛ №99 боюнча нормативдик-укуктук актылар",
+    "title2": "ПЛ №99 НУА актылары боюнча кенениреек маалымат",
+    "ustav": "КР ПРЕЗИДЕНТИНИН ЖАНЫНДАГЫ ПЛИТ №99 УЮМДУНУН УСТАВЫ",
+    "strategy": "ПЛ №99 өнүктүрүү стратегиясы 2024–2026",
+    "certificate1": "Орто билим берүү боюнча аккредитация сертификаты",
+    "certificate2": "Кесиптик билим берүү боюнча аккредитация сертификаты",
+    "certificate3": "ПЛ №99 юридикалык жактын күбөлүгү",
+    "politics": "ПЛИТ №99 билим берүүнүн сапаты боюнча саясат",
+    "mission": "Профессионалдык маалыматтык технологиялар лицейинин миссиясы №99",
+    "license1": "ПЛИТ №99 билим берүү ишин жүргүзүүгө укук берген 1620-лицензия",
+    "license2": "ПЛИТ №99 билим берүү ишин жүргүзүүгө укук берген 1854-лицензия"
   },
   "teachers": {
     "list": [
@@ -460,10 +497,6 @@
     "sports": "Спорттук курстар",
     "language": "Тил курстары",
     "professional": "Кесиптик курстар"
-  },
-  "npa": {
-    "title1": "Кыргыз Республикасынын ченемдик укуктук актылар",
-    "title2": "КР Президент алдындагы МТКЛ №99 ченемдик укуктук актылар"
   },
   "footer": {
     "logoAlt": "Лицейдин логотиби",
