+++ public/locales/RU/translation.json
@@ -169,7 +169,10 @@
     "informationSecurity": "Образовательная программа «Информационная безопасность»:",
     "miningMechanic": "Образовательная программа \"Слесарь по техническому обслуживанию и ремонту горного оборудования и спецтехники\":",
     "carRepairMechanic": "Образовательная программа \"Слесарь по ремонту автотранспортных средств\":",
-    "regulationsPlit99": "Нормативно-правовой акт: Положение ПЛИТ №99"
+    "regulationsPlit99": "Нормативно-правовой акт: Положение ПЛИТ №99",
+    "selfassessmentreports": "Отчеты по самооценке",
+    "Profession": "Образовательная программа Профессионального обучения:",
+    "normativeActsKR": "Нормативно - правовые акты Кыргызской Республики"
   },
   "docsdescriptions": {
     "webDeveloper": "Web-разработчик и мультимедийных интерфейсов",
@@ -186,6 +189,9 @@
     "miningMechanic": "Слесарь по техническому обслуживанию и ремонту горного оборудования и спецтехники - надёжность машин под землёй и на поверхности:",
     "carRepairMechanic": "Слесарь по ремонту автотранспортных средств - чтобы техника всегда была на ходу:",
     "regulationsPlit99": "Подробное Положение ПЛИТ №99 — порядок организации учебного процесса в лицее.",
+    "initvetprogramselfassessment": "По самооценке программ начального профессионального образования",
+    "Profession": "Профессиональное обучение - навыки для успешной карьеры:",
+    "ActsPLIT99": "Нормативно - правовые акты ПЛИТ №99 при Президенте КР",
     "documents": {
       "title": "Какие документы нужны для поступления",
       "text": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed doeiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enimad minim veniam, quis nostrud exercitation ullamco laboris nisi utaliquip ex ea commodo.",
@@ -258,7 +264,7 @@
     "toLessonMethodology": "Методика проведения уроков ТО",
     "toPlanTemplateScan": "Макет темплана ТО спецдисциплин (скан)",
     "poPlanTemplateScan": "Макет темплана ПО (скан)",
-    "study_programs": "Учебные программы",
+     "study_programs": "Учебные программы",
     "general_education_rp_russian_group": "РП общеобразовательный (русская группа)",
     "geometry_umk_rus": "УМК Геометрия русск.",
     "chio_rp_rus": "РП ЧиО русск.",
@@ -291,9 +297,40 @@
     "human_society_rp_kg": "РП Адам жана коом кырг.",
     "kyrgyz_language_rp_kg": "Кыргыз тили",
     "kyrgyz_language_umk_kg": "Кыргыз тили УМК",
+    "itVETprogramSelfAssessment": "По самооценке программ начального профессионального образования it направлений",
+    "geneduselfassessment": "По самооценке программы среднего общего образования ",
+    "mechvetselfassessment": "По самооценке программ начального профессионального образования по профессии слесарь по ремонту автотранспортных средств",
+    "institutionalselfassessment": "По институциональной самооценке",
     "POPD": "Рабочая программа по дисциплине «Производственное обучение»",
     "Security": "Рабочая программа по дисциплине «Охрана труда»",
-    "Profession": "Рабочая программа по дисциплине «Основа бизнеса и предпринимательства»"
+    "Profession": "Рабочая программа по дисциплине «Основа бизнеса и предпринимательства»",
+    "ActsPLIT": "Нормативный правовой акт Кыргызской Республики ПЛИТ №99",
+    "ActsExam": "Экзамен-25",
+    "PresidentialDecree": "Указ Президента КР",
+    "GovernmentDecree": "Постановление Правительство КР",
+    "CabinetOfMinisters": "Постановление Кабинета Министров КР",
+    "IECRegulations": "постановление о ВКЭ",
+    "KyrgyzStateEdDocs2025": "Положение на кырг.о документах гос.обр. 2025",
+    "StateStandardTemplate": "Макет Гос.стандарта",
+    "IGA-25": "ИГА-25",
+    "LawsOfTheKR": "Закон Кр «Об образовании»"
+  },
+  "npa": {
+    "title1": "Нормативно - правовые акты Кыргызской Республики",
+    "title2": "Нормативно - правовые акты ПЛИТ №99 при Президенте КР"
+  },
+  "npa99": {
+    "title1": "Нормативно - правовые акты ПЛ №99",
+    "title2": "Больше информации об НПА ПЛ №99",
+    "ustav": "УСТАВ УЧРЕЖДЕНИЯ ПЛИТ №99 ПРИ ПРЕЗИДЕНТЕ КР",
+    "strategy": "Стратегия развития ПЛ №99 на 2024-2026",
+    "certificate1": "Сертификат аккредитации среднее образование",
+    "certificate2": "Сертификат аккредитации профессиональное образование",
+    "certificate3": "Свидетельство юр лица ПЛ 99",
+    "politics": "Политика в области качества образования ПЛИТ №99",
+    "mission": "Миссия Профессионального лицея информационных технологий №99",
+    "license1": "Лицензия 1620 на право ведения образовательной деятельности ПЛИТ №99",
+    "license2": "Лицензия 1054 на право ведения образовательной деятельности ПЛИТ №99"
   },
   "teachers": {
     "list": [
@@ -448,10 +485,7 @@
     "language": "Языковые курсы",
     "professional": "Профессиональные курсы"
   },
-  "npa": {
-    "title1": "Нормативно - правовые акты Кыргызской Республики",
-    "title2": "Нормативно - правовые акты ПЛИТ №99 при Президенте КР"
-  },
+
   "newsPage": {
     "id1": {
       "title": "Почему исследования важны",
