+++ src/components/organisms/Header.jsx
@@ -98,13 +98,7 @@ const Header = () => {
                         >
                           {t("info.docs")}
                         </Link>
-                        <Link
-                          onClick={() => close()}
-                          to="/info/npa"
-                          className="block px-3 py-2 hover:bg-gray-400 rounded"
-                        >
-                          {t("info.npa")}
-                        </Link>
+                        
                       </div>
                     </PopoverPanel>
                   </>
@@ -254,13 +248,7 @@ const Header = () => {
                         >
                           {t("info.docs")}
                         </Link>
-                        <Link
-                          onClick={() => close()}
-                          to="/info/npa"
-                          className="block px-3 py-2 hover:bg-[#63001F] hover:text-white rounded"
-                        >
-                          {t("info.npa")}
-                        </Link>
+                        
                       </div>
                     </PopoverPanel>
                   )}
