--- src/pages/HeaderPages/NPA.async.js
@@ -1,3 +1,0 @@
-import { lazy } from "react";
-
-export const NPA = lazy(() => import("./NPA"))
