+++ src/pages/HeaderPages/Npa99.async.js
@@ -1,0 +1,3 @@
+import { lazy } from "react";
+
+export const Npa99 = lazy(() => import("./Npa99"));
