+++ src/pages/HeaderPages/NpaKr.async.js
@@ -1,0 +1,3 @@
+import { lazy } from "react";
+
+export const NpaKr = lazy(() => import("./NpaKr"));
