+++ src/pages/HeaderPages/OP.async.js
@@ -1,0 +1,3 @@
+import { lazy } from "react";
+
+export const OP = lazy(() => import("./OP"));
