+++ src/pages/HeaderPages/RPProf.async.js
@@ -1,0 +1,3 @@
+import { lazy } from "react";
+
+export const RPProf = lazy(() => import("./RPProf"));
