+++ src/pages/HeaderPages/SelfReport.async.js
@@ -1,0 +1,3 @@
+import { lazy } from "react";
+
+export const SelfReport = lazy(() => import("./SelfReport"));
