--- src/pages/HeaderPages/Teacherinfo.async.js
@@ -1,3 +1,0 @@
-import { lazy } from "react";
-
-export const Teacherinfo = lazy(() => import("./Teacherinfo"))
