+++ src/pages/Layout/Layout.jsx
@@ -34,7 +34,6 @@ const Layout = () => {
       </main>
       {isAdminPath ? null : <LocationPage />}
       {isAdminPath ? null : <Footer />}
-      <NPAPosition />
     </div>
   );
 };
